requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
scikit-learn==1.6.1
tldextract==5.1.3
requests==2.32.3
google-generativeai==0.8.5
python-whois==0.9.5
